+++ App/CMakeLists.txt
@@ -5,6 +5,7 @@ add_executable(App ClientApp.hpp main.cpp ClientApp.cpp widgets/Widget.hpp widge
 target_link_libraries(App PRIVATE Core fmt::fmt Vulkan::Vulkan)
 target_include_directories(App PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/include ${CMAKE_CURRENT_SOURCE_DIR}/widgets/include ${CMAKE_CURRENT_SOURCE_DIR}/widgets ${CMAKE_SOURCE_DIR}/ThirdParty/glm)
 target_sources(App PRIVATE version.rc)
+target_precompile_headers(App REUSE_FROM Core)
 
 file(GLOB_RECURSE VERTEX_SHADERS "${CMAKE_CURRENT_SOURCE_DIR}/shaders/*.vert")
 file(GLOB_RECURSE FRAGMENT_SHADERS "${CMAKE_CURRENT_SOURCE_DIR}/shaders/*.frag")
