+++ App/shaders/Basic.frag
@@ -3,6 +3,7 @@
 layout(set = 1, binding = 2) uniform sampler2D geometry_texture;
 
 layout(location = 0) in vec2 in_uvs;
+layout(location = 1) in vec4 shadow_pos;
 
 layout(location = 0) out vec4 colour;
