+++ App/shaders/Shadow.frag
@@ -1,0 +1,3 @@
+#version 460
+
+void main() {}
